conn: separate gso and sticky control
Android wants GSO but not sticky.
